requirements.txt
torch
torchvision
torchaudio
timm
flask
pillow
scikit-learn
matplotlib
huggingface_hub
